Changed Program.cs to use top-level entry point; removed Startup.cs
