+++ Program.cs
@@ -1,20 +1,53 @@
-﻿using System.IO;
-using Microsoft.AspNetCore;
+﻿using System.Text;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.IdentityModel.Tokens;
+using WebApi.Helpers;
+using WebApi.Services;
 
-namespace WebApi
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            BuildWebHost(args).Run();
-        }
+var builder = WebApplication.CreateBuilder(args);
 
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
-                .UseUrls("http://localhost:4000")
-                .Build();
-    }
-}
+builder.Services.AddCors();
+builder.Services.AddMvc(options => options.EnableEndpointRouting = false);
+
+// configure strongly typed settings objects
+var appSettingsSection = builder.Configuration.GetSection("AppSettings");
+builder.Services.Configure<AppSettings>(appSettingsSection);
+
+// JWT authentication
+var appSettings = appSettingsSection.Get<AppSettings>();
+var key = Encoding.ASCII.GetBytes(appSettings.Secret);
+builder.Services.AddAuthentication(options => {
+   options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+   options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+}).AddJwtBearer(options => {
+   options.RequireHttpsMetadata = false;
+   options.SaveToken = true;
+   options.TokenValidationParameters = new TokenValidationParameters() {
+      ValidateIssuerSigningKey = true,
+      IssuerSigningKey = new SymmetricSecurityKey(key),
+      ValidateIssuer = false,
+      ValidateAudience = false
+   };
+});
+
+// DI for application services
+builder.Services.AddScoped<IUserService, UserService>();
+
+// Custom url+port (ignoring launchSettings)
+builder.WebHost.UseUrls("http://localhost:4000");
+
+var app = builder.Build();
+
+app.UseCors(x => x
+   .AllowAnyOrigin()
+   .AllowAnyMethod()
+   .AllowAnyHeader());
+   
+app.UseAuthentication();
+app.UseMvc();
+
+app.Run();
